Introduce a Kerberos Login Module replacement
The Sun/Oracle JGSS implementation integrates with JAAS, allowing one
to store credentials in the Subject associated with a particular
access control context and to use those credentials. This provides
application developers with a simple way to multiplex different
identities into a Kerberized application. With respect to the MIT
Kerberos/GSSAPI implementation, this functionality depends on being
able to control the credentials cache used by the GSSAPI interfaces.
This functionality is notably lacking from the Sun "Native JGSS"
implementation (which allows use of platform-native GSSAPI libraries
on Solaris and Linux).

This patch adds a JAAS Login Module that authenticates using the
platform-native MIT Kerberos libraries and stores a handle to the
credentials cache in the access control Subject. A subsequent patch
will integrate with the JGSS implementation to use the stored
credentials.