Use existing credentials when available in Subject
When requesting credentials from the GSSManager, first check the
subject to see if matching KerberosCredentials exist; if so, import
from these instead of calling gss_acquire_creds and relying on the
default credentials cache.

This patch allows another entity (e.g. a LoginModule) to stash
credentials in a JAAS Subject for subsequent use in JGSS sessions.