Populate TGT KerberosTicket in subject
The Sun Krb5LoginModule stores that Ticket Granting Ticket for a
Kerberos-authenticated entity in the Subject. Our JGSS module doesn't
use the Subject as a credentials cache (except insofar as we stash a
handle to the underlying libkrb5 context and ccache), but existing
applications may depend on this functionality (e.g. to check directly
whether the TGT is expired before initiating GSSAPI credentials
acquisition). This patch populates a KerberosTicket structure with the
contents of the TGT and stashes them in the Subject at login time.